feat: Guixify repository
kanata-package.scm:  A definition of a package struct for Kanata which
allows it to be directly referred to in Guix System services for
example.

.guix-channel.scm:  Someone can now add this repository as a channel,
so that they can use the latest master branch of Kanata...  I'm
planning on contributing the package to upstream but it still might be
behind the development version hosted here.

guix.scm:  Allows someone to setup a development environment for
Kanata given Guix by simply invoking `guix-shell'.